Test: Revert recurrence.t test as per comment
The comment mentions one behaviour but the test was performing another.
It was also failing for a different reason, which is filters with no
matches always exit with non-zero. If the correct behaviour is indeed
expiring all recurring tasks, the test must use runError instead.

Also added a TODO for easier finding.