Test: Fix mangled "run_all --verbose" output
For lack of being able to reproduce the problem, this is a "fix by
guessing at the root cause" commit.  Don't do this at home.

FreeBSD was showing a 100.04% test suite run because the output of
"test/problems --summary" was intertwined with TAP output.

The only somewhat sane way for this to happen that does not involve
magic is that Python's sys.stdout buffer is not flushed before calling
subprocess.call(), which uses the raw file descriptors.

A quick check of the sources seems to support this theory.