New feature: urgency accumulation
Allow blocking tasks' urgency values to reflect the total urgency of the
blocked tasks, rather than just the urgency of the most urgent among the
blocked tasks.

N.B. There seems to be a bug (maybe a race condition of some sort) in
the way TaskWarrior attempts to recursively identify blocking tasks and
to get their urgency values. As a result of that bug, some "leaf" tasks
(in graph-theoretic terminology) end up being assigned the same urgency
as "branch" tasks. Until the bug is resolved, the functionality added in
this commit will not work *optimally* (nor exactly as people are likely
to expect); though it is still *useful*.

Fixes
[#2081](https://github.com/GothenburgBitFactory/taskwarrior/issues/2081).

needsTest.