Fix bash_completion test failure.
The test was failing on fresh checkout due to string mismatch in several
test cases. This was because  both new-id and new-uuid verbosity were set
during the test, so TW defaulted to printing out uuid, which did not match the
assertion.

This is very similar to commit abddb1ea8e1d15d63f4784ff33eefe80994715ca.