Properly escape filters including spaces
This adds a note to the man page to properly escape filters containing
spaces. It also fixes the unittests to reflect this.

One of the unittests contained an alternative syntax as discussed in
TW-1479 (#1505). It has been extracted into its own unittest and marked
as an expected failure because it is currently not supported.