tests: Allow match for 8 bit color code in color.rules.t
The test proper coloring for overdue task from an hour ago fails,
even though the task is properly colored. This is because the blue color
gets upgraded to blue color in 256-color scheme, and hence is
represented with 38;5;4m instead of 34m.

Support both representations of the blue color in the given test.