Bug: Patch for taskwarrior to resolve issue #2373
This patch requires a patch for libshared. This commit refers
to a libshared that includes the patch but is not in the master
branch for libshared. So this commit should not be pulled and
serves for review and testing purposes only.

After the patch for libshared has been pulled this branch can be
updated to point to the libshared on the master branch.