Replace status:waiting with +WAITING
With this change, WAITING becomes only a virtual tag, dynamically
determined based on a task's `wait` property and the current time.  This
avoids the need to modify ("un-wait") a task, something which previously
occured during gc and thus could happen on multiple replicas or might
not occur when the user expected.

The user-visible change is that filters using `status:pending` which
assume waiting tasks are not included will need to be modified to
include `-WAITING`.