colors: Do not provide any color for color.tagged by default
The simple fact that the task has a single or more tags of any value
does not convey a lot of information. As such, we should not color tasks
based on this property by default.

Users can still configure the color.tagged configuration variable if the
color rule makes sense in their system.