Use the tag-related Task methods in the Tags column type
This column type was taking a shortcut by assuming the `tags` property
contains a comma-separated list of tags. This property is now deprecated
and tags are stored in individual properties. The tag-related Task
methods correctly access these new properties.

Includes a test for the fixed behavior and a test demonstrating that
`task info` already does the right thing.