ColumnUDADuration: add "age" style which shows a more human-readable format
I just copied the logic for this from ColumnTypeDate. I also grepped around
for other instances of `formatVague`. I don't know what the argument means
but in this project it's set to `true` 100% of the time so I kept it that
way.

I am using the name "age" as suggested in
https://github.com/GothenburgBitFactory/taskwarrior/issues/3738