ColTypeDate: format "countdown" as described in example text
The "countdown" format has been the same as the "relative" format, except
that negative durations are not rendered, for about a decade. (There was
a minor fix in #1840.)

However, the example text shows something that looks like 00:00:00. This
is how "countdown" was originally rendered but this was changed as part of
the "isodate" branch in 995de68c903707a45c30e6e7b4ff124537917cf5. See
https://github.com/GothenburgBitFactory/taskwarrior/issues/4013 for more
information.

Fixes #4013