ColTypeDate: update test for "countdown" to match current code
Copy the unit test from "relative" for the positive case of "countdown",
since the code as implemented for "countdown" is identical to the code
for "relative".

I am including this commit, even though the next commit will override
it, to confirm my understanding of how the existing datetime styles
work. Initially I had gotten myself turned in circles because I misread
the "if now < date then add a - sign" logic as an absolute-value
calculation and badly misunderstood the behavior of "age" and "relative".

See https://github.com/GothenburgBitFactory/taskwarrior/issues/4013